third_party/WebKit/LayoutTests/inspector/profiler/heap-snapshot-summary-expand-collapse.html

<html>
<head>
  <script src="../../http/tests/inspector/inspector-test.js"></script>
  <script src="heap-snapshot-test.js"></script>
<script>

function test()
{
    var instanceCount = 25;
    function createHeapSnapshot()
    {
        return InspectorTest.createHeapSnapshot(instanceCount);
    }

    InspectorTest.runHeapSnapshotTestSuite([
        function testShowAll(next)
        {
            InspectorTest.takeAndOpenSnapshot(createHeapSnapshot, step1);

            function step1()
            {
                InspectorTest.switchToView("Summary", step2);
            }

            function step2()
            {
                InspectorTest.findAndExpandRow("A", step3);
            }

            function step3(row)
            {
                row.collapse();
                row.expand();
                var visibleChildren = row.children.filter(c => c._element.classList.contains('revealed'));
                InspectorTest.assertEquals(11, visibleChildren.length);
                next();
            }
        }
    ]);
}

</script>
</head>
<body onload="runTest()">
<p>
https://crbug.com/738932
Tests the snapshot view is not empty on repeatitive expand-collapse.
</p>
</body>
</html>
