third_party/WebKit/LayoutTests/inspector/profiler/heap-snapshot-summary-expand-collapse-expected.txt

https://crbug.com/738932 Tests the snapshot view is not empty on repeatitive expand-collapse.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testShowAll
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
A data-grid-data-grid-node parent revealed
undefined data-grid-data-grid-node revealed

Profiler was disabled.
